backend/requirements.txt
django-storages
boto3
sentry-sdk
django-storages
boto3
# RoadAid Network Backend

## Requirements

### Base
Django>=5.0,<5.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
drf-spectacular>=0.27.0
channels>=4.0.0
channels-redis>=4.1.0
celery>=5.3.0
redis>=5.0.0
mysqlclient>=2.2.0
python-dotenv>=1.0.0
Pillow>=10.1.0

### Geolocation
# django.contrib.gis is included in Django core
geopy>=2.4.0

### Notifications
firebase-admin>=6.2.0

### Payments
stripe>=7.0.0

### Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
factory-boy>=3.3.0

### Production
gunicorn>=21.2.0
uvicorn>=0.24.0
whitenoise>=6.6.0
